feat: improvements on quiz with/out answer on questions will still work as normal
I need to improve the UI more and add backend and also skip quiz
