+++ src/app/quiz/identification/_components/OpenEnded.tsx
@@ -2,8 +2,8 @@
 import React, { useState } from "react";
 
 interface OpenEndedProps {
-  question?: string;
-  answer?: string;
+  question: string;
+  answer: string;
   onSubmit: (userAnswer: string) => void;
 }
 
@@ -21,31 +21,67 @@ const OpenEnded: React.FC<OpenEndedProps> = ({
   const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
     onSubmit(userAnswer);
+    resetToDefaultAnswer();
   };
 
+  function resetToDefaultAnswer() {
+    setUserAnswer("");
+  }
+
+  const isAnswerPresent = question.includes(answer);
+
   return (
-    <form onSubmit={handleSubmit}>
-      {question && answer ? (
-        <>
-          {question.split(answer).map((part, index) => (
-            <React.Fragment key={index}>
-              {part}
-              {index < question.split(answer).length - 1 && (
-                <input
-                  type="text"
-                  value={userAnswer}
-                  onChange={handleChange}
-                  placeholder="Type your answer"
-                />
+    <div className="grid grid-row-2">
+      <form className="text-xl" onSubmit={handleSubmit}>
+        <div>
+          {question && answer ? (
+            <>
+              {isAnswerPresent ? (
+                <>
+                  {question.split(answer).map((part, index) => (
+                    <React.Fragment key={index}>
+                      {part}{" "}
+                      {index < question.split(answer).length - 1 && (
+                        <input
+                          className="border border-black/70 rounded-sm text-base pl-2 pr-2"
+                          type="text"
+                          value={userAnswer}
+                          onChange={handleChange}
+                          placeholder="Answer"
+                          style={{ width: "110px" }}
+                        />
+                      )}
+                    </React.Fragment>
+                  ))}
+                </>
+              ) : (
+                <p>
+                  <input
+                    className="border border-black/70 rounded-sm text-base pl-2 pr-2"
+                    type="text"
+                    value={userAnswer}
+                    onChange={handleChange}
+                    placeholder="Answer"
+                    style={{ width: "110px" }}
+                  />{" "}
+                  {question}
+                </p>
               )}
-            </React.Fragment>
-          ))}
-        </>
-      ) : (
-        <p>No quiz data available.</p>
-      )}
-      <button type="submit">Submit</button>
-    </form>
+            </>
+          ) : (
+            <p>No quiz data available.</p>
+          )}
+        </div>
+        <div className="flex items-center justify-end mt-3">
+          <button
+            className="py-1 px-5 hover:bg-blue-300/30 bg-blue-300 rounded-xl"
+            type="submit"
+          >
+            Submit
+          </button>
+        </div>
+      </form>
+    </div>
   );
 };
